Userdata mixin
