+++ auth_backend/auth_method/__init__.py
@@ -1,6 +1,8 @@
 from .base import AUTH_METHODS, AuthMethodMeta
+from .method_mixins import LoginableMixin, RegistrableMixin
 from .oauth import OauthMeta
 from .session import Session
+from .userdata_mixin import UserdataMixin
 
 
 __all__ = [
@@ -8,4 +10,7 @@ __all__ = [
     "AUTH_METHODS",
     "AuthMethodMeta",
     "OauthMeta",
+    "LoginableMixin",
+    "RegistrableMixin",
+    "UserdataMixin",
 ]
