+++ auth_backend/auth_plugins/email.py
@@ -10,9 +10,7 @@ from fastapi_sqlalchemy import db
 from pydantic import field_validator, model_validator
 from sqlalchemy import func
 
-from auth_backend.auth_method import AuthMethodMeta
-from auth_backend.auth_method import Session
-from auth_backend.auth_method.method_mixins import LoginableMixin, RegistrableMixin
+from auth_backend.auth_method import AuthMethodMeta, LoginableMixin, RegistrableMixin, Session, UserdataMixin
 from auth_backend.base import Base, StatusResponseModel
 from auth_backend.exceptions import AlreadyExists, AuthFailed, IncorrectUserAuthType, SessionExpired
 from auth_backend.kafka.kafka import get_kafka_producer
@@ -105,7 +103,7 @@ class ResetForgottenPassword(Base):
     new_password: Annotated[str, MinLen(1)]
 
 
-class Email(LoginableMixin, RegistrableMixin, AuthMethodMeta):
+class Email(UserdataMixin, LoginableMixin, RegistrableMixin, AuthMethodMeta):
     prefix = "/email"
 
     @staticmethod
