+++ auth_backend/auth_plugins/google.py
@@ -12,9 +12,7 @@ from google.auth.transport import requests
 from google.oauth2.id_token import verify_oauth2_token
 from pydantic import BaseModel, Field, Json
 
-from auth_backend.auth_method import AuthMethodMeta
-from auth_backend.auth_method import OauthMeta
-from auth_backend.auth_method import Session
+from auth_backend.auth_method import AuthMethodMeta, OauthMeta, Session
 from auth_backend.exceptions import AlreadyExists, OauthAuthFailed, OauthCredentialsIncorrect
 from auth_backend.kafka.kafka import get_kafka_producer
 from auth_backend.models.db import User, UserSession
